This commit was manufactured by cvs2svn to create tag 'CEXP_Release_2_2_2'.
Sprout from B_libtecla_orig 2007-02-14 20:34:02 UTC guest ' - importing libtecla-1.6.1'
Cherrypick from master 2013-01-28 22:29:17 UTC strauman ' - removed libtecla/Makefile (see next note) - conflicts with autotool-generated':
    Makefile.am
    Makefile.in.orig
    Makefile.rules
    README.rtems
    configure.in
    demo.c
    direader.c
    expand.c
    getline.c
    getline.c.opostdiff
    hash.c
    homedir.c
    html/cpl_complete_word.html
    html/ef_expand_file.html
    html/enhance.html
    html/gl_get_line.html
    html/gl_io_mode.html
    html/libtecla.html
    html/pca_lookup_file.html
    html/tecla.html
    keytab.c
    libtecla.h
    man/func/ef_expand_file.in
    pathutil.c
Delete:
    Makefile
    Makefile.in
    config.guess
    config.sub
    configure
    install-sh
    man3/cfc_file_start.3
    man3/cfc_literal_escapes.3
    man3/cfc_set_check_fn.3
    man3/cpl_add_completion.3
    man3/cpl_complete_word.3
    man3/cpl_file_completions.3
    man3/cpl_last_error.3
    man3/cpl_list_completions.3
    man3/cpl_record_error.3
    man3/del_CplFileConf.3
    man3/del_ExpandFile.3
    man3/del_GetLine.3
    man3/del_PathCache.3
    man3/del_PcaPathConf.3
    man3/del_WordCompletion.3
    man3/ef_expand_file.3
    man3/ef_last_error.3
    man3/ef_list_expansions.3
    man3/enhance.3
    man3/gl_change_terminal.3
    man3/gl_clear_history.3
    man3/gl_configure_getline.3
    man3/gl_customize_completion.3
    man3/gl_echo_mode.3
    man3/gl_get_line.3
    man3/gl_group_history.3
    man3/gl_ignore_signal.3
    man3/gl_last_signal.3
    man3/gl_limit_history.3
    man3/gl_load_history.3
    man3/gl_lookup_history.3
    man3/gl_prompt_style.3
    man3/gl_range_of_history.3
    man3/gl_resize_history.3
    man3/gl_save_history.3
    man3/gl_show_history.3
    man3/gl_size_of_history.3
    man3/gl_state_of_history.3
    man3/gl_terminal_size.3
    man3/gl_toggle_history.3
    man3/gl_trap_signal.3
    man3/gl_watch_fd.3
    man3/libtecla.3
    man3/libtecla_version.3
    man3/new_CplFileConf.3
    man3/new_ExpandFile.3
    man3/new_GetLine.3
    man3/new_PathCache.3
    man3/new_PcaPathConf.3
    man3/new_WordCompletion.3
    man3/pca_last_error.3
    man3/pca_lookup_file.3
    man3/pca_path_completions.3
    man3/pca_scan_path.3
    man3/pca_set_check_fn.3
    man3/ppc_file_start.3
    man3/ppc_literal_escapes.3

+++ keytab.c
@@ -575,7 +575,7 @@ static int _kt_parse_keybinding_string(const char *keyseq, char *binary,
 /*
  * Convert the character to a control character.
  */
-	*optr++ = MAKE_CTRL(c);
+	*optr++ = MAKE_CTRL((int)c);
       } else {
 	*optr++ = *iptr++;
       };
@@ -605,7 +605,7 @@ static int _kt_parse_keybinding_string(const char *keyseq, char *binary,
  */
     case 'C':
       if(_kt_is_emacs_ctrl(iptr)) {
-	*optr++ = MAKE_CTRL(iptr[2]);
+	*optr++ = MAKE_CTRL((int)iptr[2]);
 	iptr += 3;
       } else {
 	*optr++ = *iptr++;
